Warning fix: cast to expected type
Cast argument to match the expected type from the format string.

Reviewed-on: http://gerrit.openafs.org/1812
Tested-by: Derrick Brashear <shadow@dementia.org>

+++ src/afs/VNOPS/afs_vnop_flock.c
@@ -525,7 +525,7 @@ DoLockWarning(afs_ucred_t * acred)
 	lastWarnPid = pid;
 #ifdef AFS_LINUX26_ENV
 	afs_warnuser
-	    ("afs: byte-range locks only enforced for processes on this machine (pid %d (%s), user %ld).\n", pid, procname, afs_cr_uid(acred));
+	    ("afs: byte-range locks only enforced for processes on this machine (pid %d (%s), user %ld).\n", pid, procname, (long)afs_cr_uid(acred));
 #else
 	afs_warnuser
 	    ("afs: byte-range lock/unlock ignored; make sure no one else is running this program (pid %d (%s), user %ld).\n", pid, procname, afs_cr_uid(acred));
